New OOP Backend Code
At first, I created functional code to see if my simulation could handle a basic file, but I quickly learned the OOP would be much cleaner and organized, as well as future proof my project. So I completely switched the main here, and am using OOP with python completely. This is much more organized for me, and I now know that OOP is stronger in that regard.